refactor: DTO 패키지 이동, SchoolBookmarkInfoV1Response 추가
