+++ backend/src/main/java/com/festago/bookmark/dto/v1/SchoolBookmarkInfoV1Response.java
@@ -1,0 +1,14 @@
+package com.festago.bookmark.dto.v1;
+
+import com.querydsl.core.annotations.QueryProjection;
+
+public record SchoolBookmarkInfoV1Response(
+    Long id,
+    String name,
+    String logoUrl
+) {
+
+    @QueryProjection
+    public SchoolBookmarkInfoV1Response {
+    }
+}
