+++ backend/src/main/java/com/festago/bookmark/dto/v1/SchoolBookmarkV1Response.java
@@ -1,11 +1,10 @@
 package com.festago.bookmark.dto.v1;
 
-import com.festago.school.dto.v1.SchoolSearchV1Response;
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
 
 public record SchoolBookmarkV1Response(
-    SchoolSearchV1Response school,
+    SchoolBookmarkInfoV1Response school,
     LocalDateTime bookmarkCreatedAt
 ) {
