+++ backend/src/main/java/com/festago/bookmark/repository/SchoolBookmarkV1QuerydslRepository.java
@@ -5,10 +5,10 @@ import static com.festago.school.domain.QSchool.school;
 
 import com.festago.bookmark.domain.Bookmark;
 import com.festago.bookmark.domain.BookmarkType;
+import com.festago.bookmark.dto.v1.QSchoolBookmarkInfoV1Response;
 import com.festago.bookmark.dto.v1.QSchoolBookmarkV1Response;
 import com.festago.bookmark.dto.v1.SchoolBookmarkV1Response;
 import com.festago.common.querydsl.QueryDslRepositorySupport;
-import com.festago.school.dto.v1.QSchoolSearchV1Response;
 import java.util.List;
 import org.springframework.stereotype.Repository;
 
@@ -21,7 +21,13 @@ public class SchoolBookmarkV1QuerydslRepository extends QueryDslRepositorySuppor
 
     public List<SchoolBookmarkV1Response> findAllByMemberId(Long memberId) {
         return select(new QSchoolBookmarkV1Response(
-            new QSchoolSearchV1Response(school.id, school.name, school.logoUrl), bookmark.createdAt))
+            new QSchoolBookmarkInfoV1Response(
+                school.id,
+                school.name,
+                school.logoUrl
+            ),
+            bookmark.createdAt
+        ))
             .from(bookmark)
             .innerJoin(school).on(school.id.eq(bookmark.resourceId)
                 .and(bookmark.memberId.eq(memberId))
