+++ backend/src/test/java/com/festago/bookmark/presentation/v1/FestivalBookmarkV1ControllerTest.java
@@ -11,7 +11,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.festago.bookmark.application.FestivalBookmarkV1QueryService;
-import com.festago.bookmark.dto.FestivalBookmarkV1Response;
+import com.festago.bookmark.dto.v1.FestivalBookmarkV1Response;
 import com.festago.festival.dto.FestivalV1Response;
 import com.festago.festival.dto.SchoolV1Response;
 import com.festago.support.CustomWebMvcTest;
